# Improve WAL performance