This commit improve WAL performance for use case when influxdb handles
many small write requests. In old code after every write the WAL file is
synced. Sync is expensive operation that takes significant time.
In new code WAL file is synced less frequently. Once per 100ms on high
load or immediately if write load is low (there are no other waiting requests).

There is significant performance improvement tested on stress test
(BatchSize=50, WriteConcurrency=150) 2000 requests, average request size
is 2760B. Old code handles 2000 points per second, new code handles
98300 requests per second. So new code has 50x better write performance.

Closing #7222