@daviesalex 2016-11-02
I have not looked at the code, but this gets a +1 from us. Risking a few hundred ms of data in exchange for massively higher performance is a tradeoff we are certainly prepared to make.