@jwilder 2016-11-04
Are you using SSD or HDD?  I don't see the same improvement using SSDs based on the benchmarks I have in this [gist](https://gist.github.com/jwilder/e1260ac5fec90e97ffa67710eb4a356f).

```
go test -v -run non -bench BenchmarkWAL_Write_1 -benchmem ./tsdb/engine/tsm1/ > /tmp/new.txt
benchcmp /tmp/old.txt /tmp/new.txt
benchmark                  old ns/op     new ns/op     delta
BenchmarkWAL_Write_1-8     64741         62046         -4.16%

benchmark                  old allocs     new allocs     delta
BenchmarkWAL_Write_1-8     2              3              +50.00%

benchmark                  old bytes     new bytes     delta
BenchmarkWAL_Write_1-8     13            109           +738.46%
```

Also, I get a deadlock if I increase concurrency (running `BenchmarkWAL_Write_10`) benchmark.  All goriness are blocked at:

```
goroutine 73 [chan receive]:
github.com/influxdata/influxdb/tsdb/engine/tsm1.(*WAL).writeToLog(0x4206b8000, 0x532d40, 0x421009860, 0x1, 0x0, 0x0)
	/Users/jason/go/src/github.com/influxdata/influxdb/tsdb/engine/tsm1/wal.go:315 +0x7f
github.com/influxdata/influxdb/tsdb/engine/tsm1.(*WAL).WritePoints(0x4206b8000, 0x420674510, 0x1, 0x0, 0x0)
	/Users/jason/go/src/github.com/influxdata/influxdb/tsdb/engine/tsm1/wal.go:233 +0x72
github.com/influxdata/influxdb/tsdb/engine/tsm1_test.benchmarkWALWrite.func1(0x420670510, 0x4206ba140, 0x4206b8000, 0x420674510)
	/Users/jason/go/src/github.com/influxdata/influxdb/tsdb/engine/tsm1/wal_test.go:746 +0x89
created by github.com/influxdata/influxdb/tsdb/engine/tsm1_test.benchmarkWALWrite
	/Users/jason/go/src/github.com/influxdata/influxdb/tsdb/engine/tsm1/wal_test.go:750 +0x659
```
 
It looks likes the `syncWaiter` channel is never closed if another write does not come in so all pending writes will wait indefinitely.  

I think this could be fixed using a `time.Ticker` that fires every 100ms and sync if necessary.